Improvements/bugfixes for Maze, Secret Collect
Fixed bug where maze could start with a shark. This is because instead
of trying to render a 'ew' arrow, it would try to render a 'we' arrow
which doesn't exist, resulting in a random arrow. I've added a warning
just in case this happens again.

Secret collect now shows all four arrows upon failing. This might help
players understand to look for those arrows, and that they don't have to
click all four surrounding squares to try and find the arrow.