Mission strings. Block progression.
Replaced 'difficulty' with 'mission string'. The first level corresponds
to '1-1', and the last level currently corresponds to '3-3' but will
eventually be something like '10-3'. Instead of storing the furthest
difficulty the player's cleared, we now store all of the mission string
outcomes.

The player cannot progress to later worlds until they finish the third
level in a world. It is okay if they find all sharks and no frogs, they
just have to play it and reach an ending.

Level buttons are now all visible -- since the player has to clear the
third level in a world to progress, progression was permanently blocked
if those buttons weren't available