Split out 'CreatureBehavior' from 'RunningFrog'
In their current implementation, frogs always chase the cursor but
they'll eventually need to do multiple things. I've split out the
'CreatureBehavior' with a single implementation, 'FrogChaseBehavior'.
This encapsulates the behavior of chasing the cursor, and other scripts
will eventually extend CreatureBehavior with other behaviors such as
dancing.

This change only affects frogs for now -- it is still heavily in flux,
and sharks only do one thing right now. Once it is more mature or there
is a need for it, I will apply the change to sharks.

Removed unused 'old_frame' field.