End intermission timer tweak.
Instead of our ending logic waiting 4 seconds and then picking an
appropriate ending, it now picks an ending which waits 4 seconds. This
is important because we will soon have endings which do not include this
wait. Pushing this delay down further gives different endings the option
of having shorter delays or no delays at all.