Fixed bug where dancing frogs could overrun their target.
This bug resulted in frogs occasionally running too far and snapping
back into place.

The old logic used a timer so the frog would run for 8.812 seconds and
stop, and then we'd set their position to their target. Sometimes this
was off by as much as 60 pixels, resulting in an unpleasant snapping
effect.

The new logic uses the frog's position, so the frog runs until they're
within a certain radius, and until their velocity would take them
further away if they continued running. This is more accurate -- it has
theoretical bugs if the radius is too small or there are hiccups where
_process() is called with large delta values. But in practice, it is
doing what we want.