Fixed crash when combining hugging/dancing frogs
The lead frog was not given a list of dancing frogs to choreograph, but
a list of ALL frogs in the scene. This meant they'd try to bark out
dance orders to frogs only meant to hug, which confused the hugging
frogs and caused the game to crash.

Dancing frogs are now ONLY given a list of other frogs involved in their
specific dance.