Update themes and fonts for Godot 4.
These files were working fine in the editor, but exporting the game
resulted in multiple 'No loader found for resource' errors, reporting
that the .res files could not be loaded.

I opened the .theme files in the editor, removed the font, re-added the
font and saved the resources. This resulted in a lot of mysterious
changes, but now the exported resources are loaded successfully.