Comic interlude logic.
Added a comic interlude for Sunshine Shoals. This comic interlude will
play the first time a player enters the world. It will also play when
you launch the game for the first time, if you're in Sunshine Shoals.

This logic was meant to mirror Kirby's Adventure on the NES, which also
has animations like this which play at appropriate times.