Refactoring: Godot #58285, Godot #76267
Removed mention of Godot #58285 -- this issue never really applied to
our project, since 'wiggle_frames' does not have a setter anyways. The
'workaround' of requiring an explicit cast seems to be working as
intended.

About 90% of our array assignments work, and 10% of them do not. I don't
understand what the pattern is, but I've documented the remaining 10%
with references to Godot #72627.