Workaround for circular reference error in IntermissionPanel.
IntermissionPanel and BeachBallTweak reference each other, because
SceneTreeTweaks describe how to modify an IntermissionPanel and
IntermissionPanel includes a list of all scene tree tweaks. This
circular reference makes it so that BeachBallTweak (and all other
tweaks) can not be opened in the editor, because of Godot #74253
(https://github.com/godotengine/godot/issues/74253)

I've resolved the circular reference by having these scenes initialized
at runtime with a 'load' command, rather than when the scene is loaded
with a 'preload' command.