Inline most setters.
I kept around the 'set_state' setter on hand because it was being used
outside the class. Other sections of code also call 'set_state' in cases
where there's no explicit state variable. I think the existence or
absence of a 'state' variable is an implementation detail which is good
to hide with a function.