Changed export presets to zips. Linux +x flag.
Changed windows, linux export presets from .exe/.x86_64 to .zip. This
lets us get rid of our cuistom packaging code and automatically sets the
executable flag for linux systems, so that players don't need `chmod +x`
the executable before running it.

I could not change the html export to use a '.zip' -- this is not
supported, and instead exports the game into a 'frog-finder.zip.html' html file.
I changed the html export filename to 'index.html' so we no longer have to do
this in our packaging script.