Rename linux export to match old filename
The name 'frog-finder-linux-v1.26.x84_86.zip' is unnecessarily verbose
and clunky. The version number should come at the end, but doesn't
because of limitations of the Godot editor (1.26 is treated as a file
extension, resulting in files with names like '-v1.zip')

The new name is better, and preserves the old conventions as well as
conventions followed by other linux games.