Workaround for 'unique animation library' issue
When creating a new ComicPanel, the animations couldn't be edited
because they were reusing the parent AnimationPlayer's AnimationPlayer.
I've added some logic to the ComicPanel script so that when a ComicPanel
is opened in the editor, it will duplicate the parent animation library.
This lets us reuse the parent animations while still allowing changes.