Extracted frog-dance-library.tres. Godot #85689 workaround.
Because of Godot #85689 (https://github.com/godotengine/godot/issues/85689),
inherited scene AnimationPlayers cannot be edited. So for example, we
can extract a 'FrogDanceBehavior' scene which has all of our animations,
but then we can't edit them. Editing them from within the
'FrogDanceBehavior' is tedious, because we can't see the effects on the
parent sprite. Editing them from within 'RunningFrog' is impossible
because of this bug, which marks all animation tracks read-only.

To work around this, I've extracted the frog-dance-library.tres to its
own external resource. The bug does not affect external resources, only
inherited ones. I also added some code to frog-dance-library.tres so
that the external resource is saved.