Made 'very small AnimationPlayer numbers' consistent
Changed some 'very small AnimationPlayer numbers' to 0.001 instead of
using arbitrary values.

Godot #91554 (https://github.com/godotengine/godot/issues/91554)
prevents AnimationPlaybackTrack from working with a value of 0, so we
use very small numbers. Numbers that are too small (like 0.00001) often
cause the same bug, while numbers that are too big (like 0.1) result in
visible anomalies, so we've settled on 0.001. Using a consistent number
will make it easier to detect and reverse the workaround in Godot 4.3
when the bug is fixed.

Fixed an unusual y coordinate for World1ComicPage.Students1.